+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Add KeepEverythingWithMinKWords Extractor
+* Add ArticleSentence Extractor
 
 # 0.3.0 / 2017-09-12